fix: remove AlertInfoRepository abstraction and restore findByCity to PersonRepository
- Deleted AlertInfoRepository and AlertInfoInterfaceRepository to simplify architecture
- Moved findByCity(String city) method back to PersonRepository
- Updated AlertInfoService to use PersonRepository directly for city-related query
